+++ src/lib/functions/split_func.rs
@@ -1,6 +1,6 @@
 //! # split_func
 //!
-//! Split funciton which takes an environment variable name and a split by string.
+//! Split function which takes an environment variable name and a split by string.
 //!
 
 #[cfg(test)]
@@ -11,16 +11,16 @@ use crate::environment;
 
 pub(crate) fn invoke(function_args: &Vec<String>) -> Vec<String> {
     if function_args.len() != 2 {
-        error!("Split expects only 2 arguments (environment variable name, split by character)");
-        panic!("Split expects only 2 arguments (environment variable name, split by character)");
+        error!("split expects only 2 arguments (environment variable name, split by character)");
+        panic!("split expects only 2 arguments (environment variable name, split by character)");
     }
 
     let env_key = function_args[0].clone();
     let split_by = function_args[1].clone();
 
     if split_by.len() != 1 {
-        error!("Split expects a single character separator");
-        panic!("Split expects a single character separator");
+        error!("split expects a single character separator");
+        panic!("split expects a single character separator");
     }
 
     let split_by_char = split_by.chars().next().unwrap();